backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-dotenv
gspread
google-auth
pandas
python-jose[cryptography]
upstash-redis